#13474: Yolov7 model trace perf bringup
