+++ models/experimental/yolov7/runner/performant_runner_infra.py
@@ -1,0 +1,121 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC.
+
+# SPDX-License-Identifier: Apache-2.0
+
+import torch
+
+from loguru import logger
+
+import ttnn
+from ttnn.model_preprocessing import preprocess_model_parameters
+
+from models.experimental.yolov7.tt.ttnn_yolov7 import ttnn_yolov7
+from models.experimental.yolov7.reference.model import Yolov7_model
+
+from models.utility_functions import divup, is_wormhole_b0
+from tests.ttnn.utils_for_testing import assert_with_pcc
+from models.experimental.yolov7.reference.yolov7_utils import download_yolov7_weights
+from models.experimental.yolov7.ttnn_yolov7_utils import create_custom_preprocessor, load_weights
+
+
+class YOLOv7PerformanceRunnerInfra:
+    def __init__(
+        self,
+        device,
+        batch_size,
+        act_dtype,
+        weight_dtype,
+        model_location_generator=None,
+        resolution=(640, 640),
+        torch_input_tensor=None,
+    ):
+        torch.manual_seed(0)
+        self.resolution = resolution
+        self.pcc_passed = False
+        self.pcc_message = "Did you forget to call validate()?"
+        self.device = device
+        self.batch_size = batch_size
+        self.act_dtype = act_dtype
+        self.weight_dtype = weight_dtype
+        self.model_location_generator = model_location_generator
+        self.torch_input_tensor = torch_input_tensor
+        self.torch_model = Yolov7_model()
+        weights_path = "tests/ttnn/integration_tests/yolov7/yolov7.pt"
+        weights_path = download_yolov7_weights(weights_path)
+        load_weights(self.torch_model, weights_path)
+        self.torch_input_tensor = (
+            torch.randn((1, 3, 640, 640), dtype=torch.float32)
+            if self.torch_input_tensor is None
+            else self.torch_input_tensor
+        )
+
+        self.parameters = preprocess_model_parameters(
+            initialize_model=lambda: self.torch_model, custom_preprocessor=create_custom_preprocessor(None), device=None
+        )
+
+        nx_ny = [80, 40, 20]
+        grid_tensors = []
+        for i in range(3):
+            yv, xv = torch.meshgrid([torch.arange(nx_ny[i]), torch.arange(nx_ny[i])])
+            grid_tensors.append(torch.stack((xv, yv), 2).view((1, 1, nx_ny[i], nx_ny[i], 2)).float())
+
+        self.ttnn_yolov7_model = ttnn_yolov7(self.device, self.parameters, grid_tensors)
+        self.torch_output_tensor = self.torch_model(self.torch_input_tensor)[0]
+
+    def _setup_l1_sharded_input(self, device, torch_input_tensor=None):
+        if is_wormhole_b0():
+            core_grid = ttnn.CoreGrid(y=8, x=8)
+        else:
+            exit("Unsupported device")
+
+        num_devices = device.get_num_devices()
+        torch_input_tensor = self.torch_input_tensor if torch_input_tensor is None else torch_input_tensor
+
+        n, c, h, w = torch_input_tensor.shape
+        ## Converting from image based channels (3) to min channels (16)
+        if c == 3:
+            c = 16
+        input_mem_config = ttnn.create_sharded_memory_config(
+            [n, c, h, w],
+            ttnn.CoreGrid(x=8, y=8),
+            ttnn.ShardStrategy.HEIGHT,
+        )
+        tt_inputs_host = ttnn.from_torch(torch_input_tensor, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT)
+
+        return tt_inputs_host, input_mem_config
+
+    def setup_dram_sharded_input(self, device, torch_input_tensor=None, mesh_mapper=None, mesh_composer=None):
+        tt_inputs_host, input_mem_config = self._setup_l1_sharded_input(device)
+        dram_grid_size = device.dram_grid_size()
+        dram_shard_spec = ttnn.ShardSpec(
+            ttnn.CoreRangeSet(
+                {ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(dram_grid_size.x - 1, dram_grid_size.y - 1))}
+            ),
+            [
+                divup(tt_inputs_host.volume() // tt_inputs_host.shape[-1], dram_grid_size.x * dram_grid_size.y),
+                tt_inputs_host.shape[-1],
+            ],
+            ttnn.ShardOrientation.ROW_MAJOR,
+        )
+        sharded_mem_config_DRAM = ttnn.MemoryConfig(
+            ttnn.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.BufferType.DRAM, dram_shard_spec
+        )
+
+        return tt_inputs_host, sharded_mem_config_DRAM, input_mem_config
+
+    def run(self):
+        self.output_tensor = self.ttnn_yolov7_model(self.input_tensor)[0]
+
+    def validate(self, output_tensor=None, torch_output_tensor=None):
+        ttnn_output_tensor = self.output_tensor if output_tensor is None else output_tensor
+        torch_output_tensor = self.torch_output_tensor if torch_output_tensor is None else torch_output_tensor
+        output_tensor = ttnn.to_torch(ttnn_output_tensor)
+
+        self.pcc_passed, self.pcc_message = assert_with_pcc(self.torch_output_tensor, output_tensor, pcc=0.99)
+
+        logger.info(
+            f"Yolov7 - batch_size={self.batch_size}, act_dtype={self.act_dtype}, weight_dtype={self.weight_dtype}, PCC={self.pcc_message}"
+        )
+
+    def dealloc_output(self):
+        ttnn.deallocate(self.output_tensor)
